only dispose renderer if initialized

+++ src/renderers/WebGLRenderer.ts
@@ -240,6 +240,8 @@ export class WebGLRenderer {
         };
 
         this.dispose = () => {
+            if (!initialized) return;
+
             worker.terminate();
 
             gl.deleteShader(vertexShader);
